backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
prisma==0.11.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment and Configuration
python-dotenv==1.0.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Data Processing and Analysis
pandas==2.1.4
numpy==1.24.3

# Financial Data APIs
yfinance==0.2.28
ccxt==4.1.64
alpha-vantage==2.3.1

# AI and Machine Learning (basic)
openai==1.3.7
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2

# Technical Analysis
pandas-ta==0.3.14b0

# Caching and Background Tasks
redis==5.0.1
celery==5.3.4

# WebSocket support
websockets==12.0

# CORS
fastapi-cors==0.0.6

# Validation
email-validator==2.1.0